+++ src/api/table.js
@@ -1,7 +1,7 @@
-import fetch from '@/utils/fetch'
+import request from '@/utils/request'
 
 export function getList(params) {
-  return fetch({
+  return request({
     url: '/table/list',
     method: 'get',
     params
